ISSUE #3888 - refactor components names
